Extracted test_suite_usercases
